rrd-graph/analysis/pyproject.toml
[tool.poetry]
name = "analysis"
version = "0.1.0"
description = ""
authors = ["CC <46365490+cc4i@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
uvicorn = "^0.32.0"
fastapi = "^0.115.4"
pytz = "^2024.2"
pg8000 = "^1.31.2"
sqlalchemy = "^2.0.36"
langgraph = "^0.2.44"
langchain-google-genai = "^2.0.4"
langchain-google-vertexai = "^2.0.7"
pydantic = "^2.9.2"
google-cloud-run = "^0.10.10"
google-cloud-storage = "^2.18.2"
google-cloud-language = "^2.15.0"
google-cloud-aiplatform = "^1.71.1"
google-auth = "^2.35.0"
google-cloud-alloydb-connector = "^1.4.0"
python-dotenv = "^1.0.1"
anthropic = "^0.39.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
